core-service: add ArticleConverter Service

+++ Back-End/core-service/src/main/java/com/aperture/community/core/module/param/UmsArticleParam.java
@@ -25,7 +25,7 @@ public class UmsArticleParam {
 
     private String content;
 
-    private Integer circleId;
+    private Long circleId;
 
     private List<Integer> tags;
